IGNITE-25279 Merge two different ZonePartitionReplicaListenerTest
